FH_DEV
Update data types of transaction model

+++ src/main/java/org/feedhandler/models/Transaction.java
@@ -6,14 +6,14 @@ public class Transaction {
 	
 	private Integer channeldId;
 	private Integer millisecondTimestamp;
-	private Double amount;
-	private Double price;
+	private long amount;
+	private long price;
 	
 	public Transaction(HashMap<String, String> message) {
 		this.channeldId = Integer.parseInt(message.get("channelId"));
 		this.millisecondTimestamp = Integer.parseInt(message.get("mts"));
-		this.amount = Double.parseDouble(message.get("mts"));
-		this.price = Double.parseDouble(message.get("price"));
+		this.amount = Long.parseLong(message.get("mts"));
+		this.price = Long.parseLong(message.get("price"));
 	}
 	
 	public String toString() {
